# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DucklingDuckling": {
        "properties": {
          "Caps": {
            "$ref": "#/components/schemas/DucklingCapabilities"
          },
          "Cost": {
            "$ref": "#/components/schemas/ConfigCost"
          },
          "ID": {
            "type": "string"
          },
          "Model": {
            "type": "string"
          },
          "Notes": {
            "type": "string"
          },
          "Params": {
            "$ref": "#/components/schemas/ConfigSamplingParams"
          },
          "Provider": {
            "type": "string"
          },
          "Roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "Caps",
          "Cost",
          "ID",
          "Model",
          "Notes",
          "Params",
          "Provider",
          "Roles"
        ],
        "type": "object"
      },
...
      "EngineapiducklingTestRequest": {
        "properties": {
          "prompt": {
            "type": "string"
          },
          "stream": {
            "type": "boolean"
          }
        },
        "required": [
          "prompt",
          "stream"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local engine API. Loopback only, bearer-token authenticated (07).",
    "title": "ducklab engine",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/ducklings": {
      "get": {
        "operationId": "DucklingList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/DucklingDuckling"
                      },
                      "type": "array"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List ducklings"
      }
    },
...
    "/v1/ducklings/{id}/test": {
      "post": {
        "operationId": "DucklingTest",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngineapiducklingTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineapiducklingTestResponse"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "One round-trip against a duckling"
      }
    },
...
    "/v1/events": {
      "get": {
        "operationId": "get_events",
        "responses": {
          "204": {
            "description": "no content"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "SSE event stream. Accepts ?token= because EventSource cannot set headers."
      }
    },
...
    "/v1/projects": {
      "get": {
        "operationId": "ProjectList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/ServiceProject"
                      },
                      "type": "array"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List registered projects"
      },
      "post": {
        "operationId": "ProjectInit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProject"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Open or initialise a project"
      }
    },
...
    "/v1/projects/{id}/roster": {
      "get": {
        "operationId": "RosterGet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRosterView"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Resolved roster"
      },
      "put": {
        "operationId": "RosterSet",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngineapirosterSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRosterView"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Assign a duckling to a role"
      }
    },
    "/v1/projects/{id}/roster/suggest": {
      "get": {
        "operationId": "RosterSuggest",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/ServiceSuggestion"
                      },
                      "type": "array"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Ranked roster suggestion"
      },
      "post": {
        "operationId": "RosterApply",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRosterView"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Apply the suggestion"
      }
    },
    "/v1/projects/{id}/runs": {
      "post": {
        "operationId": "RunStart",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunlogRun"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Start a run"
      }
    },
    "/v1/projects/{id}/status": {
      "get": {
        "operationId": "ProjectStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceStatus"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Project status"
      }
    },
...
    "/v1/runs": {
      "get": {
        "operationId": "RunList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/RunlogRun"
                      },
                      "type": "array"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List runs"
      }
    },
...
    "/v1/runs/{id}/abort": {
      "post": {
        "operationId": "RunAbort",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "no content"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Abort"
      }
    },
...
    "/v1/runs/{id}/answer": {
      "post": {
        "operationId": "RunAnswer",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngineapianswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "no content"
          },
          "401": {
            "description": "missing or invalid credentials"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Answer a pending question"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "variables": {
        "port": {
          "default": "0"
        }
      }
    }
  ]
}